+++ src/components/ProjectName/ProjectName.test.js
@@ -53,7 +53,7 @@ describe("With a webComponent=true", () => {
     store = mockStore(initialState);
     render(
       <Provider store={store}>
-        <ProjectName showLabel={true} forWebComponent={true} />
+        <ProjectName showLabel={true} editable={false} />
       </Provider>,
     );
   });